src/main/java/stringmanipulation/alternating/Solution.java
package stringmanipulation.alternating;

import java.io.BufferedWriter;
import java.io.FileWriter;
import java.io.IOException;
import java.util.Scanner;

public class Solution {

    // Complete the alternatingCharacters function below.
    static int alternatingCharacters(String s) {
        final int[] result = { 0 };
        final String[] previous = { "" };

        s
                .codePoints()
                .mapToObj(i -> String.valueOf((char) i))
                .forEach(cStr -> {
                    if (cStr.equals(previous[0])) {
                        result[0]++;
                        return;
                    }

                    previous[0] = cStr;
                });

        return result[0];
    }

    private static final Scanner scanner = new Scanner(System.in);

    public static void main(String[] args) throws IOException {
        BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(System.getenv("OUTPUT_PATH")));

        int q = scanner.nextInt();
        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");

        for (int qItr = 0; qItr < q; qItr++) {
            String s = scanner.nextLine();

            int result = alternatingCharacters(s);

            bufferedWriter.write(String.valueOf(result));
            bufferedWriter.newLine();
        }

        bufferedWriter.close();

        scanner.close();
    }
}
